Moved more functionality to the buffer files
