+++ transfer/fabtsuite_buffer.c
@@ -9,7 +9,10 @@
 #include <stdio.h>
 #include <stdlib.h>
 
+#include <rdma/fi_domain.h>
+
 #include "fabtsuite_buffer.h"
+#include "fabtsuite_error.h"
 #include "fabtsuite_types.h"
 #include "fabtsuite_util.h"
 
@@ -101,6 +104,12 @@ progbuf_alloc(void)
     return pb;
 }
 
+bool
+progbuf_is_wellformed(progbuf_t *pb)
+{
+    return pb->hdr.nused == sizeof(pb->msg);
+}
+
 vecbuf_t *
 vecbuf_alloc(void)
 {
@@ -150,4 +159,93 @@ vecbuf_is_wellformed(vecbuf_t *vb)
         return true;
 
     return false;
+}
+
+/*
+ * Registration functions
+ */
+
+int
+buf_mr_bind(bufhdr_t *h, struct fid_ep *ep)
+{
+    int rc;
+
+    if (ep == NULL)
+        h->ep = NULL;
+    else if (global_state.mr_endpoint &&
+             (rc = fi_mr_bind(h->mr, &ep->fid, 0)) != 0)
+        return rc;
+    else if (global_state.mr_endpoint && (rc = fi_mr_enable(h->mr)) != 0)
+        return rc;
+    else
+        h->ep = ep;
+
+    h->desc = fi_mr_desc(h->mr);
+
+    return 0;
+}
+
+int
+buf_mr_dereg(bufhdr_t *h)
+{
+    struct fid_mr *mr = h->mr;
+
+    if (mr == NULL)
+        return 0;
+
+    h->mr = NULL;
+    return fi_close(&mr->fid);
+}
+
+int
+buf_mr_reg(struct fid_domain *dom, struct fid_ep *ep, uint64_t access,
+           uint64_t key, bufhdr_t *h)
+{
+    int rc;
+    bytebuf_t *b = (bytebuf_t *) h;
+
+    rc = fi_mr_reg(dom, &b->payload[0], h->nallocated, access, 0, key, 0,
+                   &h->mr, NULL);
+
+    if (rc != 0) {
+        h->mr = NULL;
+        return rc;
+    }
+
+    if ((rc = buf_mr_bind(h, ep)) != 0) {
+        (void) buf_mr_dereg(h);
+        return rc;
+    }
+
+    return 0;
+}
+
+bufhdr_t *
+progbuf_create_and_register(struct fid_ep *ep)
+{
+    progbuf_t *pb = progbuf_alloc();
+    int rc;
+
+    rc = buf_mr_reg(global_state.domain, ep, FI_SEND,
+                    seqsource_get(&global_state.keys), &pb->hdr);
+
+    if (rc != 0)
+        bailout_for_ofi_ret(rc, "buf_mr_reg");
+
+    return &pb->hdr;
+}
+
+bufhdr_t *
+vecbuf_create_and_register(struct fid_ep *ep)
+{
+    vecbuf_t *vb = vecbuf_alloc();
+    int rc;
+
+    rc = buf_mr_reg(global_state.domain, ep, FI_SEND,
+                    seqsource_get(&global_state.keys), &vb->hdr);
+
+    if (rc != 0)
+        bailout_for_ofi_ret(rc, "buf_mr_reg");
+
+    return &vb->hdr;
 }
